osd: fix signed/unsigned comp
warning: osd/ReplicatedPG.cc:1208: comparison between signed and unsigned integer expressions

+++ src/osd/ReplicatedPG.cc
@@ -1205,7 +1205,7 @@ int ReplicatedPG::do_osd_ops(OpContext *ctx, vector<OSDOp>& ops,
 	  struct stat st;
 	  memset(&st, 0, sizeof(st));
 	  result = osd->store->stat(coll, soid, &st);
-	  if (result || st.st_size != oi.size) {
+	  if (result || (uint64_t)st.st_size != oi.size) {
 	    osd->clog.error() << info.pgid << " " << soid << " oi.size " << oi.size
 			      << " but stat got " << result << " size " << st.st_size << "\n";
 	    assert(0 == "oi disagrees with stat");
